# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FleetOps SADC API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect after login.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to OIDC logout.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  # ─── Org ────────────────────────────────────────────────────────────────
...
  /compliance-docs:
    get:
      tags: [Compliance]
      operationId: listComplianceDocs
      summary: List compliance documents for the org
      parameters:
        - name: entityType
          in: query
          required: false
          schema:
            type: string
        - name: entityId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplianceDoc'
    post:
      tags: [Compliance]
      operationId: createComplianceDoc
      summary: Add a compliance document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceDocInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceDoc'
  /compliance-docs/{docId}:
    get:
      tags: [Compliance]
      operationId: getComplianceDoc
      summary: Get a single compliance document
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceDoc'
    patch:
      tags: [Compliance]
      operationId: updateComplianceDoc
      summary: Update a compliance document
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceDocUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceDoc'
    delete:
      tags: [Compliance]
      operationId: deleteComplianceDoc
      summary: Delete a compliance document
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted

  # ─── Fuel Purchases ────────────────────────────────────────────────────────
...
  /border-catalog:
    get:
      tags: [BorderCatalog]
      operationId: listBorderCharges
      summary: List border charge catalog entries
      parameters:
        - name: borderPost
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BorderCharge'
    post:
      tags: [BorderCatalog]
      operationId: createBorderCharge
      summary: Add a border charge catalog entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorderChargeInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorderCharge'
  /border-catalog/{chargeId}:
    patch:
      tags: [BorderCatalog]
      operationId: updateBorderCharge
      summary: Update a border charge
      parameters:
        - name: chargeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorderChargeUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorderCharge'
    delete:
      tags: [BorderCatalog]
      operationId: deleteBorderCharge
      summary: Delete a border charge entry
      parameters:
        - name: chargeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted

  # ─── Object Storage ────────────────────────────────────────────────────────
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true

    # ─── Auth Schemas ─────────────────────────────────────────────────────
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string

    # ─── Org Schemas ──────────────────────────────────────────────────────
...
    ComplianceDoc:
      type: object
      required: [id, orgId, entityType, entityId, docType, expiryDate, docStatus, createdAt]
      properties:
        id:
          type: string
        orgId:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        docType:
          type: string
        docNumber:
          type: [string, 'null']
        issuer:
          type: [string, 'null']
        country:
          type: [string, 'null']
        issueDate:
          type: [string, 'null']
          format: date
        expiryDate:
          type: string
          format: date
        fileUrl:
          type: [string, 'null']
        docStatus:
          type: string
        daysUntilExpiry:
          type: [integer, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: [string, 'null']
          format: date-time
    ComplianceDocInput:
      type: object
      required: [entityType, entityId, docType, expiryDate]
      properties:
        entityType:
          type: string
        entityId:
          type: string
        docType:
          type: string
        docNumber:
          type: string
        issuer:
          type: string
        country:
          type: string
        issueDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        fileUrl:
          type: string
    ComplianceDocUpdate:
      type: object
      properties:
        docNumber:
          type: string
        issuer:
          type: string
        country:
          type: string
        issueDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        fileUrl:
          type: string

    # ─── Fuel Schemas ─────────────────────────────────────────────────────
...
    BorderCharge:
      type: object
      required: [id, borderPost, country, chargeType, defaultAmount, currency, appliesTo]
      properties:
        id:
          type: string
        borderPost:
          type: string
        country:
          type: string
        chargeType:
          type: string
        defaultAmount:
          type: number
        currency:
          type: string
        appliesTo:
          type: string
    BorderChargeInput:
      type: object
      required: [borderPost, country, chargeType, defaultAmount, currency, appliesTo]
      properties:
        borderPost:
          type: string
        country:
          type: string
        chargeType:
          type: string
        defaultAmount:
          type: number
        currency:
          type: string
        appliesTo:
          type: string
    BorderChargeUpdate:
      type: object
      properties:
        defaultAmount:
          type: number
        currency:
          type: string
        appliesTo:
          type: string

    # ─── Storage Schemas ──────────────────────────────────────────────────
...