# lib/api-zod/src/generated/types/borderChargeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FleetOps SADC API
 * OpenAPI spec version: 0.1.0
 */

export interface BorderChargeUpdate {
  defaultAmount?: number;
  currency?: string;
  appliesTo?: string;
}